# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.EventType:
    enum:
    - wedding
    - birthday
    - corporate
    type: string
    x-enum-varnames:
    - EventTypeWedding
    - EventTypeBirthday
    - EventTypeCorporate
  models.Menu:
    properties:
      id:
        type: integer
      img:
        type: string
      name_kz:
        type: string
      name_ru:
        type: string
      restaurant_id:
        type: integer
    type: object
...
  models.Restaurant:
    properties:
      _2gis_map:
        type: string
      address_kz:
        type: string
      address_ru:
        type: string
      city_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
  models.RestaurantEvent:
    properties:
      desc:
        type: string
      eventtype:
        $ref: '#/definitions/models.EventType'
      id:
        type: integer
      img:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  models.Section:
    properties:
      id:
        type: integer
      name:
        type: string
      restaurant_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API сервера для системы управления рестораном
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management System API
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Возвращает список всех событий ресторана
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RestaurantEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить список всех событий ресторана
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Создает новое событие ресторана в системе
      parameters:
      - description: Данные события ресторана
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.RestaurantEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новое событие ресторана
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет событие ресторана по его ID
      parameters:
      - description: ID события ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить событие ресторана
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Возвращает событие ресторана по его ID
      parameters:
      - description: ID события ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RestaurantEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить событие ресторана по ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего события ресторана
      parameters:
      - description: ID события ресторана
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные события ресторана
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.RestaurantEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные события ресторана
      tags:
      - events
  /events/type/{type}:
    get:
      consumes:
      - application/json
      description: Возвращает список событий ресторана указанного типа
      parameters:
      - description: Тип события (wedding, birthday, corporate)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RestaurantEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить события ресторана по типу
      tags:
      - events
...
  /menus/restaurant/{restaurantID}:
    get:
      consumes:
      - application/json
      description: Возвращает список меню для указанного ресторана
      parameters:
      - description: ID ресторана
        in: path
        name: restaurantID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить меню по ID ресторана
      tags:
      - menus
  /restaurants:
    get:
      consumes:
      - application/json
      description: Возвращает список ресторанов с фильтрацией по активности
      parameters:
      - default: true
        description: Фильтр по активности ресторанов
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Restaurant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить список ресторанов
      tags:
      - restaurants
    post:
      consumes:
      - application/json
      description: Создает новый ресторан в системе
      parameters:
      - description: Данные ресторана
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/models.Restaurant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новый ресторан
      tags:
      - restaurants
  /restaurants/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет ресторан по его ID
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить ресторан
      tags:
      - restaurants
    get:
      consumes:
      - application/json
      description: Возвращает ресторан по его ID
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Restaurant'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить ресторан по ID
      tags:
      - restaurants
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего ресторана
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные ресторана
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/models.Restaurant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные ресторана
      tags:
      - restaurants
...
  /sections/restaurant/{restaurantID}:
    get:
      consumes:
      - application/json
      description: Возвращает список секций для указанного ресторана
      parameters:
      - description: ID ресторана
        in: path
        name: restaurantID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Section'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить секции по ID ресторана
      tags:
      - sections
...
swagger: "2.0"